init profile tree
